move file handler to own module
